+++ Nowtes/nbproject/project.properties
@@ -36,13 +36,15 @@ dist.jlink.output=${dist.jlink.dir}/Nowtes
 endorsed.classpath=
 excludes=
 file.reference.jcalendar-1.4.jar=C:\\conectorjavamsq\\jcalendar-1.4.jar
+file.reference.jcalendar-1.4.jar-1=C:\\Users\\PC\\Downloads\\jcalendar-1.4\\lib\\jcalendar-1.4.jar
 file.reference.mysql-connector-java-8.0.26.jar=C:\\conectorjavamsq\\mysql-connector-java-8.0.26.jar
 includes=**
 jar.compress=true
 javac.classpath=\
     ${file.reference.mysql-connector-java-8.0.26.jar}:\
     ${libs.absolutelayout.classpath}:\
-    ${file.reference.jcalendar-1.4.jar}
+    ${file.reference.jcalendar-1.4.jar}:\
+    ${file.reference.jcalendar-1.4.jar-1}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
