--- Nowtes/src/test/MainNowtes.java
@@ -1,26 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package test;
-
-import view.MainView;
-
-/**
- *
- * @author PC
- */
-public class MainNowtes {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-
-        MainView mainview = new MainView();
-        mainview.setVisible(true);
-        
-    }
-    
-}
