+++ Nowtes/src/test/Test.java
@@ -4,9 +4,7 @@
  * and open the template in the editor.
  */
 package test;
-
-import MainView.MainView;
-
+import view.MainView;
 /**
  *
  * @author PC GOOSE
@@ -17,8 +15,7 @@ public class Test {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        // TODO code application logic here
-        MainView vistaPrincipal = new MainView();
+       MainView xd = new MainView();
         
        
     }
