+++ Nowtes/src/view/AlertDialog/AlertDialog.form
@@ -75,9 +75,6 @@
         <Property name="font" type="java.awt.Font" editor="org.netbeans.beaninfo.editors.FontEditor">
           <Font name="Trebuchet MS" size="14" style="0"/>
         </Property>
-        <Property name="icon" type="javax.swing.Icon" editor="org.netbeans.modules.form.editors2.IconEditor">
-          <Image iconType="2" name="C:\Users\polil\OneDrive\Escritorio\xddddddddd.png"/>
-        </Property>
         <Property name="opaque" type="boolean" value="true"/>
       </Properties>
     </Component>
