+++ Nowtes/src/view/AlertDialog/AlertDialog.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package MainView;
+package view.AlertDialog;
 
 /**
  *
