+++ Nowtes/src/view/EditNote/EditNote.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package MainView;
+package view.EditNote;
 
 import com.toedter.calendar.JDateChooser;
 import java.awt.event.ActionEvent;
