+++ Nowtes/src/view/ViewCompleteNote/ViewCompleteNote.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package MainView;
+package view.CompletedNote;
 
 
 import java.awt.event.ActionEvent;
